moves configuration unit test to different file
